Disable removing driver when relationship exict

+++ resources/views/driver/driverslist.blade.php
@@ -67,7 +67,7 @@
                             @endcomponent
                             {{-- <a class="badge badge-secondary  mb-1 ml-1" href="/driver/show/{{$driver->id}}">Info &#x27BE;</a> --}}
                         @endif
-                        @if (Auth::user()->role_id > 3 )
+                        @if (Auth::user()->role_id > 3 && $driver->services->count()==0 && $driver->fillups->count()==0 && $driver->rides->count()==0)
                             <a class="badge badge-danger mb-1 ml-1" href="/driver/delete/{{$driver->id}}"><i class="material-icons md-16">person_add_disabled</i></a>
                         @endif
                     </div>
